Merge v2.6.6 to master

+++ xabber/src/main/java/com/xabber/android/data/push/PushManager.java
@@ -5,7 +5,6 @@ import android.content.Context;
 import android.provider.Settings;
 import android.util.Log;
 
-import com.google.firebase.iid.FirebaseInstanceId;
 import com.xabber.android.BuildConfig;
 import com.xabber.android.data.Application;
 import com.xabber.android.data.SettingsManager;
@@ -20,13 +19,14 @@ import com.xabber.android.data.entity.AccountJid;
 import com.xabber.android.data.entity.UserJid;
 import com.xabber.android.data.http.PushApiClient;
 import com.xabber.android.data.log.LogManager;
+import com.xabber.android.utils.ExternalAPIs;
 import com.xabber.android.utils.Utils;
 
 import org.jivesoftware.smack.SmackException;
 import org.jivesoftware.smack.XMPPException;
 import org.jivesoftware.smack.packet.IQ;
 import org.jivesoftware.smack.packet.Stanza;
-import org.jivesoftware.smack.tcp.XMPPTCPConnection;
+import com.xabber.xmpp.smack.XMPPTCPConnection;
 import org.jivesoftware.smackx.disco.ServiceDiscoveryManager;
 import org.jivesoftware.smackx.push_notifications.element.EnablePushNotificationsIQ;
 import org.jivesoftware.smackx.push_notifications.element.PushNotificationsElements;
@@ -219,9 +219,12 @@ public class PushManager implements OnConnectedListener, OnPacketListener {
     /** Private */
 
     private void registerEndpoint(AccountJid accountJid) {
+        String token = ExternalAPIs.getPushEndpointToken();
+        if (token == null) return;
+
         compositeSubscription.add(
                 PushApiClient.registerEndpoint(
-                        FirebaseInstanceId.getInstance().getToken(),
+                        token,
                         accountJid.getFullJid().asBareJid().toString() + getAndroidId())
                         .subscribeOn(Schedulers.io())
                         .observeOn(AndroidSchedulers.mainThread())
@@ -239,9 +242,12 @@ public class PushManager implements OnConnectedListener, OnPacketListener {
     }
 
     private void deleteEndpoint(final AccountItem accountItem) {
+        String token = ExternalAPIs.getPushEndpointToken();
+        if (token == null) return;
+
         compositeSubscription.add(
                 PushApiClient.deleteEndpoint(
-                        FirebaseInstanceId.getInstance().getToken(), accountItem.getAccount().toString())
+                        token, accountItem.getAccount().toString())
                         .subscribeOn(Schedulers.io())
                         .observeOn(AndroidSchedulers.mainThread())
                         .subscribe(new Action1<ResponseBody>() {
